MB-23857 issue warnings when rows are filtered out of system tables
Note that these warnings can be issued multiple times per query,
but the user should only see one warning per query. Hence the code in
service_response.go to filter out duplicates.

Reviewed-on: http://review.couchbase.org/77662
Tested-by: Johan Larson <johan.larson@couchbase.com>

+++ server/http/service_response.go
@@ -347,11 +347,16 @@ loop:
 func (this *httpRequest) writeWarnings(prefix, indent string) bool {
 	var err errors.Error
 	ok := true
+	alreadySeen := make(map[string]bool)
 loop:
 	for ok {
 		select {
 		case err, ok = <-this.Warnings():
 			if ok {
+				if err.OnceOnly() && alreadySeen[err.Error()] {
+					// do nothing for this warning
+					continue loop
+				}
 				if this.warningCount == 0 {
 					this.writeString(",\n")
 					this.writeString(prefix)
@@ -359,6 +364,7 @@ loop:
 				}
 				ok = this.writeError(err, this.warningCount, prefix, indent)
 				this.warningCount++
+				alreadySeen[err.Error()] = true
 			}
 		default:
 			break loop
